--- src/apps/an-delay/hooks/useNow.js
@@ -1,14 +1,0 @@
-import { useEffect, useState } from 'react'
-
-export function useNow(updateEvery = 1000) {
-  const [now, setNow] = useState(new Date())
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setNow(new Date())
-    }, updateEvery)
-    return () => {
-      clearInterval(timer)
-    }
-  }, [updateEvery])
-  return now
-}
